Move validation of a table format in a label from TableDataContentValidationRule to TableFieldDefinitionRule:
1. Remove validation code from TableDataContentValidationRule.java
2. Remove unused imports statements from TableDataContentValidationRule.java
3. Add validation code to TableFieldDefinitionRule.java
4. Add TableFieldDefinitionRule to validation-commands.xml so TableFieldDefinitionRule can be executed on a label
5. Update problem type from TABLE_FILE_READ_ERROR to INVALID_LABEL when reporting error to match reported error from TableFieldDefinitionRule.java in ValidationIntegrationTests.java

Refs #11 https://github.com/nasa-pds/validate/issues/11 Update allowable field_format values per Standards Reference definition regarding [+|-] characters
