+++ src/main/resources/validation-commands.xml
@@ -16,6 +16,7 @@
     <command caption="Validate Referenced Local Identifiers in label" className="gov.nasa.pds.tools.validate.rule.pds4.LocalIdentifierReferencesRule"/>
     <command caption="Register file references" className="gov.nasa.pds.tools.validate.rule.pds4.RegisterTargetReferences" />
     <command caption="Register label identifiers" className="gov.nasa.pds.tools.validate.rule.pds4.RegisterLabelIdentifiers" />
+    <command caption="Data Structure matches table descriptions in label" className="gov.nasa.pds.tools.validate.rule.pds4.TableFieldDefinitionRule" />
     <command caption="Data Content matches table descriptions in label" className="gov.nasa.pds.tools.validate.rule.pds4.TableDataContentValidationRule" />
     <command caption="Data Content matches array descriptions in label" className="gov.nasa.pds.tools.validate.rule.pds4.ArrayContentValidationRule" />
     <command caption="Record validation results" className="gov.nasa.pds.tools.validate.rule.RecordValidationResults" />
